added basic function class to hold all data related to functions

+++ src/main.cpp
@@ -23,23 +23,33 @@ int main(int argc, char* argv[])
     }
 
     std::vector<Token> tokens = tokenize_file(argv[1]);
-    std::map<std::string, std::vector<Op>> program = parse_tokens(tokens);
+    std::map<std::string, Function> program = parse_tokens(tokens);
+
+    ASSEMBLER assembler = FASM;
 
     // compile
     std::cout << INFO_COLOR << "[info] generating asm from '" << argv[1] << "'." << RESET_COLOR << std::endl;
-    compile_to_asm(program, "out.asm");
+    compile_to_asm(program, "out.asm", assembler);
 
-    // nasm
-    std::cout << INFO_COLOR << "[info] generating object file." << RESET_COLOR << std::endl;
-    exit_on_error(std::system("nasm -felf64 out.asm"));
+    // nasm or fasm
+    std::cout << INFO_COLOR << "[info] compiling assembly." << RESET_COLOR << std::endl;
 
-    // link
-    std::cout << INFO_COLOR << "[info] linking object file." << RESET_COLOR << std::endl;
-    exit_on_error(std::system("ld -o out out.o"));
+    if (assembler == FASM)
+    {
+        exit_on_error(std::system("fasm out.asm"));
+    }
+    else
+    {
+        exit_on_error(std::system("nasm -felf64 out.asm"));
+
+        // link
+        std::cout << INFO_COLOR << "[info] linking object file." << RESET_COLOR << std::endl;
+        exit_on_error(std::system("ld -o out out.o"));
     
-    // cleanup
-    std::cout << INFO_COLOR << "[info] cleaning up object file" << RESET_COLOR << std::endl;
-    exit_on_error(std::system("rm out.o"));
+        // cleanup
+        std::cout << INFO_COLOR << "[info] cleaning up object file" << RESET_COLOR << std::endl;
+        exit_on_error(std::system("rm out.o"));
+    }
 
     return 0;
 }
